requirements.txt
numpy
pillow
scikit-image
matplotlib
pymodm
flask
flask-cors
pytest
pytest-pep8
pytest-cov
